refac frontedlogging package
